refactor(web): login/signup の共通送信フローを useAuthSubmit へ集約
use-login と use-signup はバリデーション・CAPTCHA確認・送信・エラー整形が
ほぼ同一だったため、その定型フローを entities/auth の useAuthSubmit にまとめ、
各フックはエンドポイント・エラー解決・成功後処理だけを注入する薄いラッパーにする。
挙動は不変で、既存の use-login/use-signup テストは無改変で通過する。
