+++ application/apps/web/src/client/entities/auth/index.ts
@@ -1,4 +1,5 @@
 export { AUTH_ERROR_MESSAGES, resolveCommonAuthErrorMessage } from './lib/error-message'
 export { resolveTurnstileSiteKey } from './lib/turnstile'
+export { default as useAuthSubmit } from './model/use-auth-submit'
 export { default as useSession, SESSION_SWR_KEY } from './model/use-session'
 export { type AuthenticateErrors, authenticateFormSchema } from './model/validation'
